requirements.txt
# LangChain et ses dépendances
langchain
langchain-openai
langchain_community

# Gestion des documents
pypdf
docx2txt
pdfminer.six

# Base de données vectorielle
chromadb
sentence-transformers

# API et modèles
openai
pydantic

# Interface utilisateur
streamlit

# Utilitaires
python-dotenv
requests

# FastAPI et uvicorn
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Multipart
python-multipart

